requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
gunicorn==20.1.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib